AshPyro.Helpers (AshPyro v0.2.1)

Shared helpers used to implement your own AshPyro components.

Functions

ash_class(fun, assigns)

AshPyro DSL allows component classes to be functions accepting assigns. This helper function simplifies handling that case in components:

<:col
  :for={col <- display_columns(@config.columns, @display)}
  class={ash_class(col.class, col)}>